requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
breadability==0.1.20
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
joblib==1.2.0
lxml==4.9.2
lyricsgenius==3.0.1
nltk==3.8.1
numpy==1.24.2
Pillow==9.5.0
pycountry==22.3.5
python-dotenv==1.0.0
pytz==2022.7.1
regex==2023.3.23
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
sumy==0.11.0
textblob==0.17.1
tqdm==4.65.0
urllib3==1.26.15
